ProjectZero/src/main/java/com/company/apps/Item.java
package com.company.apps;

import java.math.BigDecimal;

public class Item {
    private int item_id=0;
    private String item_name= "";
    private BigDecimal item_price=BigDecimal.valueOf(0);
    private int quantity= 0;

    public int getItem_id() {
        return item_id;
    }
    public void setItem_id(int item_id) {
        this.item_id = item_id;
    }

    public String getItem_name() {
        return item_name;
    }

    public void setItem_name(String item_name) {
        this.item_name = item_name;
    }

    public BigDecimal getItem_price() {
        return item_price;
    }

    public void setItem_price(BigDecimal item_price) {
        this.item_price = item_price;
    }

    public int getQuantity() {
        return quantity;
    }

    public void setQuantity(int quantity) {
        this.quantity = quantity;
    }

}
